Function serde_dynamodb::from_hashmap

pub fn from_hashmap<'a, T, S: BuildHasher + Clone>(
    hm: HashMap<String, AttributeValue, S>
) -> Result<T> where
    T: Deserialize<'a>, 

Deserialize an instance of type T from an HashMap<String, AttributeValue>.

Errors

This conversion can fail if the structure of the input does not match the structure expected by T, for example if T is a struct type but the input contains something other than a JSON map. It can also fail if the structure is correct but T's implementation of Deserialize decides that something is wrong with the data, for example required struct fields are missing from the JSON map or some number is too big to fit in the expected primitive type.